requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
chardet==4.0.0
charset-normalizer==2.1.0
discord-components==2.1.2
discord-py-interactions==4.2.1
discord-py-slash-command==4.2.1
discord.py==0.16.12
DiscordUtils==1.3.4
frozenlist==1.3.0
idna==3.3
multidict==5.2.0
py-cord==1.7.3
typing_extensions==4.2.0
websockets==3.4
yarl==1.7.2